Various tweaks

+++ Source/Engine/Tools/ModelTool/ModelTool.cpp
@@ -1431,7 +1431,7 @@ bool ModelTool::ImportModel(const String& path, ModelData& data, Options& option
                             Transform srcNode = data.Skeleton.Nodes[nodeIndex].LocalTransform;
                             auto& node = data.Skeleton.Nodes[nodeIndex];
                             if (auto* channel = animation.GetChannel(node.Name))
-                                channel->Evaluate(frame, &srcNode, false);
+                                channel->Evaluate((float)frame, &srcNode, false);
                             pose.Nodes[nodeIndex] = srcNode;
                         }
 
@@ -1439,7 +1439,7 @@ bool ModelTool::ImportModel(const String& path, ModelData& data, Options& option
                         key = Float3::Zero;
                         for (int32 nodeIndex = 0; nodeIndex < nodes; nodeIndex++)
                             key += pose.GetNodeModelTransformation(data.Skeleton, nodeIndex).Translation;
-                        key /= nodes;
+                        key /= (float)nodes;
                     }
 
                     // Calculate skeleton center of mass movement over the animation frames
@@ -1448,7 +1448,7 @@ bool ModelTool::ImportModel(const String& path, ModelData& data, Options& option
                     for (int32 frame = 0; frame < frames; frame++)
                     {
                         auto& key = rootChannel.Position[frame];
-                        key.Time = frame;
+                        key.Time = (float)frame;
                         key.Value = centerOfMass[frame] - centerOfMassRefPose;
                     }
 
@@ -1494,7 +1494,7 @@ bool ModelTool::ImportModel(const String& path, ModelData& data, Options& option
                                 Transform srcNode = data.Skeleton.Nodes[nodeIndex].LocalTransform;
                                 auto& node = data.Skeleton.Nodes[nodeIndex];
                                 if (auto* channel = animation.GetChannel(node.Name))
-                                    channel->Evaluate(frame, &srcNode, false);
+                                    channel->Evaluate((float)frame, &srcNode, false);
                                 pose.Nodes[nodeIndex] = srcNode;
                             }
